# backend/internal/controller/transport/http/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Emfont Backend API",
    "version": "v1",
    "description": "Emfont font generation API backed by PostgreSQL, MinIO, and HarfBuzz."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "paths": {
...
    "/openapi.json": {
      "get": {
        "summary": "OpenAPI document",
        "operationId": "getOpenAPI",
        "responses": {
          "200": {
            "description": "OpenAPI JSON document."
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            }
          },
          "500": {
            "description": "OpenAPI document generation failed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/docs": {
      "get": {
        "summary": "API documentation",
        "operationId": "getDocs",
        "responses": {
          "200": {
            "description": "HTML API documentation."
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ErrorDetail": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "INVALID_VALUE",
              "INVALID_FORMAT",
              "INVALID_ENUM",
              "REQUIRED",
              "VALUE_TOO_SHORT",
              "VALUE_TOO_LONG",
              "VALUE_TOO_SMALL",
              "VALUE_TOO_LARGE"
            ]
          },
          "message": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "value": {}
        }
      },
      "Problem": {
        "type": "object",
        "required": [
          "code",
          "title",
          "status",
          "detail"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "detail": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            }
          }
        }
      },
...
    }
  }
}